+++ fmt/Omnius.Xeus.Service.Implements/Omnius.Xeus.Service.Internal.rpd
@@ -4,16 +4,10 @@ using "Omnius.Core.Cryptography";
 using "Omnius.Core.Network";
 using "Omnius.Xeus.Service";
 
-namespace "Omnius.Xeus.Service.Components.Internal";
+namespace "Omnius.Xeus.Service.Internal";
 option csharp_access_level = "internal";
 
 table MerkleTreeSection {
     Length: uint64 = 0,
     Hashes: vector<OmniHash>(capacity: 1073741824) = 1,
-}
-
-table TcpConnectorConfig {
-    Version: uint32 = 0,
-    TcpConnectOptions: TcpConnectOptions? = 1,
-    TcpAcceptOptions: TcpAcceptOptions? = 2,
 }
