+++ src/Omnius.Xeus.Service.Implements/Omnius.Xeus.Service.Implements.csproj
@@ -18,4 +18,8 @@
     <ProjectReference Include="..\Omnius.Xeus.Service\Omnius.Xeus.Service.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="Primitives\" />
+  </ItemGroup>
+
 </Project>
