--- test/Omnius.Xeus.Service.Tests/FileStorageTests.cs
@@ -1,9 +1,0 @@
-using Omnius.Xeus.Service.Primitives;
-
-namespace Omnius.Xeus.Service
-{
-    public class FileStorageTests : TestsBase
-    {
-    }
-}
-
